Add makefile, Move mock to internal
